requirements.txt
pandas>=1.3.0
numpy>=1.21.0
openpyxl>=3.0.0
PyYAML>=5.4.0
streamlit>=1.20.0
plotly>=5.0.0
requests>=2.25.0
pytest>=7.0.0
